cli: Add a sub-command to upload curations to ClearlyDefined
For an example of a created ClearlyDefined pull-request see e.g. [1].

Note that only components that already exist on ClearlyDefined can be
curated, and this command will fail with

    "The contribution has failed because some of the supplied component
    definitions do not exist"

if a curation refers to a non-existing components. This will be fixed
later by using the harvest API [2] to get the component created first.

[1] https://github.com/clearlydefined/curated-data-dev/pull/942
[2] https://api.clearlydefined.io/api-docs/#/harvest/post_harvest

Signed-off-by: Sebastian Schuberth <sebastian.schuberth@bosch-si.com>